A class to manage BAM files.

This class will allow to load, convert and normalize alignments and regions files/data.

Bam_Handler

Format

A BAM manager

Value

Bam_Handler$new returns a Bam_Handler object which contains coverage related information for every BAM files.

Constructor

bh <- Bam_Handler$new(bam_files, cores = SerialParam())

bam_files

A vector of BAM filenames. The BAM files must be indexed. i.e.: if a file is named file.bam, there must be a file named file.bam.bai or file.bai in the same directory.

cores

The number of cores available to parallelize the analysis. Either a positive integer or a BiocParallelParam. Default: SerialParam().

paired_end

If TRUE, metagene will deal with paired-end data. If FALSE, single-end data are expected

Bam_Handler$new returns a Bam_Handler object that contains and manages BAM files. Coverage related information as alignment count can be obtain by using this object.

Methods

bh$get_aligned_count(bam_file)

bam_file

The name of the BAM file.

bg$get_bam_name(bam_file)

bam_file

The name of the BAM file.

bh$get_rpm_coefficient(bam_file)

bam_file

The name of the BAM file.

bh$index_bam_files(bam_files)

bam_files

A vector of BAM filenames.

bh$get_bam_files()

bh$get_coverage(bam_file, regions) force_seqlevels = FALSE)

bam_file

The name of the BAM file.

regions

A not empty GRanges object.

force_seqlevels

If TRUE, Remove regions that are not found in bam file header. Default: FALSE. TRUE and FALSE respectively correspond to pruning.mode = "coarse" and "error" in ?seqinfo.

bh$get_normalized_coverage(bam_file, regions) force_seqlevels = FALSE)

bam_file

The name of the BAM file.

regions

A not empty GRanges object.

force_seqlevels

If TRUE, Remove regions that are not found in bam file header. Default: FALSE. TRUE and FALSE respectively correspond to pruning.mode = "coarse" and "error" in ?seqinfo.

bh$get_noise_ratio(chip_bam_file, input_bam_file)

chip_bam_file

The path to the chip bam file.

input_bam_file

The path to the input (control) bam file.

Examples

bam_file <- get_demo_bam_files()[1]
bh <- metagene2:::Bam_Handler$new(bam_files = bam_file)
bh$get_aligned_count(bam_file)

#> [1] 1128